requirements.txt
fastapi
uvicorn
pydantic
nump
pandas
scikit-learn==1.3.2
joblib
langgraph
langchain-openai
langchain-core
openai
python-multipart
python-dotenv
imblearn